requirements.txt
pyyaml
zeep>=3.0.0,<4.0.0
